pyproject.toml
[project]
name = "codex-sangpye-skill"
version = "0.1.0"
description = "Korean e-commerce 상세페이지(상폐) generator — 13 sections + 1080x7500 combined image via Codex OAuth. No OpenAI API key required."
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "genie"}]
keywords = ["codex", "image-generation", "ecommerce", "korean", "sangpye", "detail-page", "skill"]
dependencies = [
    "Pillow>=11.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
sangpye = "sangpye_skill.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sangpye_skill"]

[tool.pytest.ini_options]
markers = [
    "live: hits real codex OAuth (skipped by default; opt in with `pytest -m live`)",
]
addopts = "-m 'not live'"
